+++ crates/rtlola-frontend/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implemented `Display` for Mir to display the whole specification
 - `ty()` and `name()` methods for Stream
 - Added `dependency_graph()` to generate dependency graph as json or in dot-format
+- `Window` trait for shared information of sliding and discrete window
 
 ## [0.5.3] - 12.04.2022